\name{RNCBIAxis2Libs}
\alias{RNCBIAxis2Libs}
\title{Apache Axis2 libraries}
\description{Provides selected Apache Axis2 libraries for the use with the RNCBI package. For license information see the inst/java directory.}
\section{Warning}{
This package has no user accessible methods.
}
\author{Martin Schumann (\email{martin.schumann.email@googlemail.com})}
